requirements.txt
python-telegram-bot==13.15
playwright==1.43.0
